[KYV-759] Created abstract class for functional tests that includes required test annotations and utility methods.
